+++ packages/deepseek/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ai-sdk/deepseek
 
+## 1.0.0-beta.11
+
+### Patch Changes
+
+- Updated dependencies [dd5fd43]
+  - @ai-sdk/provider-utils@3.0.0-beta.8
+  - @ai-sdk/openai-compatible@1.0.0-beta.11
+
 ## 1.0.0-beta.10
 
 ### Patch Changes