+++ packages/fal/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ai-sdk/fal
 
+## 1.0.0-beta.8
+
+### Patch Changes
+
+- eb173f1: chore (providers): remove model shorthand deprecation warnings
+- Updated dependencies [dd5fd43]
+  - @ai-sdk/provider-utils@3.0.0-beta.8
+
 ## 1.0.0-beta.7
 
 ### Patch Changes