+++ packages/llamaindex/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @ai-sdk/llamaindex
 
+## 1.0.0-beta.31
+
+### Patch Changes
+
+- Updated dependencies [6622441]
+- Updated dependencies [ced8eee]
+- Updated dependencies [cee64b2]
+- Updated dependencies [ee38081]
+- Updated dependencies [dd5fd43]
+  - ai@5.0.0-beta.31
+
+## 1.0.0-beta.30
+
+### Patch Changes
+
+- ai@5.0.0-beta.30
+
 ## 1.0.0-beta.29
 
 ### Patch Changes