+++ packages/openai-compatible/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/openai-compatible
 
+## 1.0.0-beta.11
+
+### Patch Changes
+
+- Updated dependencies [dd5fd43]
+  - @ai-sdk/provider-utils@3.0.0-beta.8
+
 ## 1.0.0-beta.10
 
 ### Patch Changes