+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ai-sdk/vue
 
+## 2.0.0-beta.32
+
+### Patch Changes
+
+- Updated dependencies [6622441]
+- Updated dependencies [ced8eee]
+- Updated dependencies [cee64b2]
+- Updated dependencies [ee38081]
+- Updated dependencies [dd5fd43]
+  - ai@5.0.0-beta.31
+  - @ai-sdk/provider-utils@3.0.0-beta.8
+
+## 2.0.0-beta.31
+
+### Patch Changes
+
+- ai@5.0.0-beta.30
+
 ## 2.0.0-beta.30
 
 ### Patch Changes